[C-1922] Fix share encoding
