+++ packages/common/src/utils/formatUtil.ts
@@ -218,3 +218,15 @@ export const formatMessageDate = (date: string) => {
   if (d.isBefore(today, 'day')) return d.format('dddd h:mm A')
   return d.format('h:mm A')
 }
+
+/**
+ * Generate a short base36 hash for a given string.
+ * Used to generate short hashes for for queries and urls.
+ */
+export const getHash = (str: string) =>
+  Math.abs(
+    str.split('').reduce((a, b) => {
+      a = (a << 5) - a + b.charCodeAt(0)
+      return a & a
+    }, 0)
+  ).toString(36)
