+++ packages/web/src/components/collectibles/components/CollectibleDetails.tsx
@@ -5,7 +5,8 @@ import {
   Collectible,
   CollectibleMediaType,
   profilePageSelectors,
-  collectibleDetailsUIActions
+  collectibleDetailsUIActions,
+  getHash
 } from '@audius/common'
 import { LogoEth, LogoSol } from '@audius/stems'
 import cn from 'classnames'
@@ -17,7 +18,6 @@ import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import PerspectiveCard from 'components/perspective-card/PerspectiveCard'
 import PreloadImage from 'components/preload-image/PreloadImage'
 import { preload } from 'utils/image'
-import { getHash } from 'utils/route'
 
 import { getFrameFromGif } from '../ethCollectibleHelpers'
