+++ packages/web/src/pages/collectibles-playlist-page/CollectiblesPlaylistPageProvider.tsx
@@ -25,7 +25,8 @@ import {
   QueueSource,
   collectibleDetailsUIActions,
   shareModalUIActions,
-  playerSelectors
+  playerSelectors,
+  getHash
 } from '@audius/common'
 import cn from 'classnames'
 import { push } from 'connected-react-router'
@@ -37,7 +38,7 @@ import { AUDIO_NFT_PLAYLIST } from 'common/store/smart-collection/smartCollectio
 import { TablePlayButton } from 'components/table/components/TablePlayButton'
 import { getLocationPathname } from 'store/routing/selectors'
 import { AppState } from 'store/types'
-import { getHash, AUDIO_NFT_PLAYLIST_PAGE, profilePage } from 'utils/route'
+import { AUDIO_NFT_PLAYLIST_PAGE, profilePage } from 'utils/route'
 
 import { CollectionPageProps as DesktopCollectionPageProps } from '../collection-page/components/desktop/CollectionPage'
 import { CollectionPageProps as MobileCollectionPageProps } from '../collection-page/components/mobile/CollectionPage'
