+++ packages/web/src/utils/route.ts
@@ -1,19 +1,7 @@
-import { ID, encodeUrlName } from '@audius/common'
+import { ID, encodeUrlName, getHash } from '@audius/common'
 import { push as pushRoute } from 'connected-react-router'
 import { Location as HistoryLocation } from 'history'
 import { matchPath } from 'react-router'
-
-/**
- * Generate a short base36 hash for a given string.
- * Used to generate short hashes for for queries and urls.
- */
-export const getHash = (str: string) =>
-  Math.abs(
-    str.split('').reduce((a, b) => {
-      a = (a << 5) - a + b.charCodeAt(0)
-      return a & a
-    }, 0)
-  ).toString(36)
 
 const USE_HASH_ROUTING = process.env.REACT_APP_USE_HASH_ROUTING === 'true'
